rename xmc
