--- transport/internet/finalmask/minecraft/config.proto
@@ -1,16 +1,0 @@
-syntax = "proto3";
-
-package xray.transport.internet.finalmask.minecraft;
-option csharp_namespace = "Xray.Transport.Internet.Finalmask.Minecraft";
-option go_package = "github.com/xtls/xray-core/transport/internet/finalmask/minecraft";
-option java_package = "com.xray.transport.internet.finalmask.minecraft";
-option java_multiple_files = true;
-
-
-message Config {
-    string password = 1;
-    repeated string usernames = 2;
-    bytes rsa_private_key = 8;
-    bytes rsa_public_key = 9;
-    string hostname = 10;
-}
