+++ transport/internet/finalmask/xmc/client.go
@@ -1,4 +1,4 @@
-package minecraft
+package xmc
 
 import (
 	"bufio"
