+++ transport/internet/finalmask/xmc/config.go
@@ -1,4 +1,4 @@
-package minecraft
+package xmc
 
 import (
 	"fmt"
