+++ transport/internet/finalmask/xmc/derivation.go
@@ -1,4 +1,4 @@
-package minecraft
+package xmc
 
 import (
 	"crypto/rsa"
