+++ transport/internet/finalmask/xmc/derivation_test.go
@@ -1,4 +1,4 @@
-package minecraft
+package xmc
 
 import (
 	"crypto/sha256"
