+++ transport/internet/finalmask/xmc/handshake_test.go
@@ -1,4 +1,4 @@
-package minecraft
+package xmc
 
 import (
 	"bytes"
