+++ transport/internet/finalmask/xmc/server.go
@@ -1,4 +1,4 @@
-package minecraft
+package xmc
 
 import (
 	"bufio"
